Move masking logic to its own sub-module.
