+++ src/skeleton/mod.rs
@@ -1,6 +1,7 @@
+mod version_masking;
+
 use crate::OptimisationProfile;
 use anyhow::Context;
-use cargo_manifest::Value;
 use fs_err as fs;
 use globwalk::{GlobWalkerBuilder, WalkError};
 use serde::{Deserialize, Serialize};
@@ -21,7 +22,7 @@ pub struct Manifest {
     pub contents: String,
 }
 
-struct ParsedManifest {
+pub(in crate::skeleton) struct ParsedManifest {
     relative_path: PathBuf,
     contents: toml::Value,
 }
@@ -34,7 +35,7 @@ impl Skeleton {
         let mut lock_file = read_lockfile(&base_path)?;
         let config_file = read_config(&base_path)?;
 
-        mask_local_crate_versions(&mut manifests, &mut lock_file);
+        version_masking::mask_local_crate_versions(&mut manifests, &mut lock_file);
 
         let lock_file = lock_file.map(|l| {
             toml::to_string(&l)
@@ -418,89 +419,3 @@ fn read_lockfile<P: AsRef<Path>>(
         }
     }
 }
-
-
-/// All local dependencies are emptied out when running `prepare`.
-/// We do not want the recipe file to change if the only difference with
-/// the previous docker build attempt is the version of a local crate
-/// encoded in `Cargo.lock` (while the remote dependency tree
-/// is unchanged) or in the corresponding `Cargo.toml` manifest.
-/// We replace versions of local crates in `Cargo.lock` and in all `Cargo.toml`s, including
-/// when specified as dependency of another crate in the workspace.
-fn mask_local_crate_versions(mut manifests: &mut [ParsedManifest], mut lock_file: &mut Option<Value>) {
-    fn mask_local_versions_in_lockfile(lock_file: &mut toml::Value, local_package_names: &[Value]) {
-        if let Some(packages) = lock_file
-            .get_mut("package")
-            .and_then(|packages| packages.as_array_mut())
-        {
-            packages
-                .iter_mut()
-                // Find all local crates
-                .filter(|package| {
-                    package
-                        .get("name")
-                        .map(|name| local_package_names.contains(name))
-                        .unwrap_or_default()
-                })
-                // Mask the version
-                .for_each(|package| {
-                    if let Some(version) = package.get_mut("version") {
-                        *version = toml::Value::String(CONST_VERSION.to_string())
-                    }
-                });
-        }
-    }
-
-    fn mask_local_versions_in_manifests(manifests: &mut [ParsedManifest], local_package_names: &[Value]) {
-        for manifest in manifests.iter_mut() {
-            if let Some(package) = manifest.contents.get_mut("package") {
-                if let Some(version) = package.get_mut("version") {
-                    *version = toml::Value::String(CONST_VERSION.to_string());
-                }
-            }
-            mask_local_dependency_versions(local_package_names, manifest, "dependencies");
-            mask_local_dependency_versions(local_package_names, manifest, "dev-dependencies");
-            mask_local_dependency_versions(local_package_names, manifest, "build-dependencies");
-        }
-    }
-
-    fn mask_local_dependency_versions(
-        local_package_names: &[Value],
-        manifest: &mut ParsedManifest,
-        dependency_key: &str,
-    ) {
-        if let Some(dependencies) = manifest.contents.get_mut(dependency_key) {
-            for local_package in local_package_names.iter() {
-                if let toml::Value::String(local_package) = local_package {
-                    let local_package = local_package.replace("-", "_");
-                    if let Some(local_dependency) = dependencies.get_mut(local_package) {
-                        if let Some(version) = local_dependency.get_mut("version") {
-                            *version = toml::Value::String(CONST_VERSION.to_string());
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    fn parse_local_crate_names(manifests: &[ParsedManifest]) -> Vec<toml::Value> {
-        let mut local_package_names = vec![];
-        for manifest in manifests.iter() {
-            if let Some(package) = manifest.contents.get("package") {
-                if let Some(name) = package.get("name") {
-                    local_package_names.push(name.to_owned());
-                }
-            }
-        }
-        local_package_names
-    }
-
-    const CONST_VERSION: &str = "0.0.1";
-
-    let local_package_names = parse_local_crate_names(&manifests);
-    mask_local_versions_in_manifests(&mut manifests, &local_package_names);
-    if let Some(l) = &mut lock_file {
-        mask_local_versions_in_lockfile(l, &local_package_names);
-    }
-}
-
